Create Order integration

+++ src/main/java/com/client/api/ws/rasmooplus/dto/wsraspay/OrderDTO.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -11,6 +13,6 @@ public class OrderDTO {
 
     private String id;
     private String customerId;
-    private Long discount;
+    private BigDecimal discount;
     private String productAcronym;
 }
